x11/gtk3: Add an expressed option for xdamage support; set PKGREVISION
GTK+3 provides a build option --enable-xdamage/--disable-xdamage

This changeset adds the symbol 'xdamage' to the PKG_OPTIONS.gtk3 list,
to permit for expressly setting that enab/disab feature, in the build.
When that feature is enabled, the build will now have a formal
dependency on x11/libXdamage.

This may serve to resolve the following error, such that may occur on
some hosts, during the configure in x11/gtk3

  checking for GTK... no
  configure: error: Package requirements (
    gtk+-3.0 >= 3.20.0
    glib-2.0 >= 2.53.2
    gmodule-2.0
    libxml-2.0 >= 2.4.0
  ) were not met:

  Package 'xdamage', required by 'gdk-3.0', not found

  Consider adjusting the PKG_CONFIG_PATH environment variable if you
  installed software in a non-standard prefix.

  Alternatively, you may set the environment variables GTK_CFLAGS
  and GTK_LIBS to avoid the need to call pkg-config.
  See the pkg-config man page for more details.
  *** Error code 1

  Stop.
  bmake[1]: stopped in /usr/pkgsrc/devel/glade

This changeset also adds a PKGREVISION to the GTK+3 Makefile - beginning
at the numerical index, '2', in this changeset.