x11/gtk3: Merge libXdamage linkage into x11 support
In maintaining some alignment with x11 platform support in MesaLib, add
libXdamage linkage for all x11 builds.

This changeset represents an effective continuation of changesets
fcb9008ada9 and 0dfc3693020 at thinkum-sys pkgsrc contrib. The original
rationale for this series of changesets is described in the first
changelog entry, for changeset fcb9008ada9

This changeset removes the xdamage build option -- such that was added,
tentatively, in the previous changeset in this series, i.e 0dfc3693020

The port x11/libXdamage will now comprise a dependency for all GTK+3
builds in which the x11 pkg option is enabled, in this local ports
tree. This dependency will also be inherited for any ports on an x11
platform, as such, linking againt x11/gtk3 as via the corresponding
buildlink mk-file.

This is in maintaining some alignment with the x11 platform support
provided in graphics/MesaLib

(cherry picked from commit 24329a3f5cae1557f4e659cb4c0f0ef292ba4a66)