devel/apr: Add a formal jlibtool build option
- Move the jlibtool subst into a conditional eval in options.mk

- Add a new build option, jlibtool

- Ensure that the bytecode jlibtool from devel/apr will not be
  overwritten previous to the build, and will not be overwritten during
  install, when the pkg is built with the jlibtool build option

- Update patch and subst strings for setting the SHELL path in
  jlibtool.c constants, in addition to setting the AR and RANLIB paths
  at time of build

This port may now be built as to provide a jlibtool installation,
without using packages under wip. Albeit, this jlibtool itself may be
somewhat more out-of-date than that available in wip/jlibtool

When this pkg is built without the jlibtool option (the default), the
file libexec/apr/libtool will nonetheless be available under LOCALBASE,
albeit as - in effect - providing a copy of the libtool used during the
build.