*: Remove esound support from pkgsrc
Enlightened Sound Daemon was one of the earlier solutions to the old
"multiple programs can't open /dev/audio at once" problem that was once
a thing we had to worry about.

Eventually, it was adopted as part of GNOME. GNOME lost interest in it
about a decade ago and dropped it in favour of PulseAudio, newer
applications are generally uninterested in supporting it. Last release
was in 2008 and support for newer OS APIs is pretty nonexistent.

Several years ago the original website disappeared.

https://en.wikipedia.org/wiki/Enlightened_Sound_Daemon
https://tracker.debian.org/news/999428/removed-0241-11-from-unstable/