bootstrap/bootstrap: Use MKDIR via an optional environment variable
When bootstrapping pkgsrc with Linux emulation on FreeBSD, the behaviors
of the mkdir command on the build host may differ to the behaviors of
the mkdir command under the linuxulator prefix. When run outside of a
chroot environment, the linux mkdir command on the build host may create
a directory at a pathname relative to the path denoted in the sysctl mib
compat.linux.emul_path on FreeBSD

This patch for the bootstrap script allows for specifying an exact
mkdir cmd to use in mkdir_p_early, such that the provided mkdir
command may not be what the script would resolve from PATH in the
build environment.